Update Trilinos after a long time

+++ packages/kokkos-kernels/src/common/KokkosKernels_Handle.hpp
@@ -41,9 +41,11 @@
 //@HEADER
 */
 #include <KokkosKernels_ExecSpaceUtils.hpp>
-#include "KokkosGraph_GraphColorHandle.hpp"
+#include "KokkosGraph_Distance1ColorHandle.hpp"
+#include "KokkosGraph_Distance2ColorHandle.hpp"
 #include "KokkosSparse_gauss_seidel_handle.hpp"
 #include "KokkosSparse_spgemm_handle.hpp"
+#include "KokkosSparse_spadd_handle.hpp"
 #ifndef _KOKKOSKERNELHANDLE_HPP
 #define _KOKKOSKERNELHANDLE_HPP
 
@@ -51,56 +53,124 @@ namespace KokkosKernels{
 
 namespace Experimental{
 
-template <class lno_row_view_t_, class lno_nnz_view_t_, class scalar_nnz_view_t_,
+template <class size_type_, class lno_t_, class scalar_t_,
           class ExecutionSpace, class TemporaryMemorySpace, class PersistentMemorySpace>
-class KokkosKernelsHandle{
+class KokkosKernelsHandle
+{
 public:
 
-  typedef ExecutionSpace HandleExecSpace;
-  typedef TemporaryMemorySpace HandleTempMemorySpace;
-  typedef PersistentMemorySpace HandlePersistentMemorySpace;
+  typedef typename ExecutionSpace::execution_space HandleExecSpace;
+  typedef typename TemporaryMemorySpace::memory_space HandleTempMemorySpace;
+  typedef typename PersistentMemorySpace::memory_space HandlePersistentMemorySpace;
+  //typedef Kokkos::Device<ExecutionSpace::execution_space,ExecutionSpace::memory_space>  HandleExecSpace;
+  //typedef Kokkos::Device<TemporaryMemorySpace::execution_space,TemporaryMemorySpace::memory_space> HandleTempMemorySpace;
+  //typedef Kokkos::Device<PersistentMemorySpace::execution_space,PersistentMemorySpace::memory_space> HandlePersistentMemorySpace;
 
-  typedef lno_row_view_t_ in_lno_row_view_t;
-  typedef lno_nnz_view_t_ in_lno_nnz_view_t;
-  typedef scalar_nnz_view_t_ in_scalar_nnz_view_t;
+  typedef typename std::remove_const<size_type_>::type  size_type;
+  typedef const size_type const_size_type;
 
-  typedef typename in_lno_row_view_t::non_const_value_type size_type;
-  typedef typename in_lno_row_view_t::array_layout row_lno_view_array_layout;
-  typedef typename in_lno_row_view_t::device_type row_lno_view_device_t;
-  typedef typename in_lno_row_view_t::memory_traits row_lno_view_memory_traits;
-  typedef typename in_lno_row_view_t::HostMirror row_lno_host_view_t; //Host view type
-  typedef typename in_lno_nnz_view_t::non_const_value_type nnz_lno_t;
-  typedef typename in_lno_nnz_view_t::array_layout nnz_lno_view_array_layout;
-  typedef typename in_lno_nnz_view_t::device_type nnz_lno_view_device_t;
-  typedef typename in_lno_nnz_view_t::memory_traits nnz_lno_view_memory_traits;
-  typedef typename in_lno_nnz_view_t::HostMirror nnz_lno_host_view_t; //Host view type
-  typedef typename in_scalar_nnz_view_t::non_const_value_type nnz_scalar_t;
-  typedef typename in_scalar_nnz_view_t::array_layout nnz_scalar_view_array_layout;
-  typedef typename in_scalar_nnz_view_t::device_type nnz_scalar_view_device_t;
-  typedef typename in_scalar_nnz_view_t::memory_traits nnz_scalar_view_memory_traits;
-  typedef typename in_scalar_nnz_view_t::HostMirror nnz_scalar_view_t; //Host view type
-  typedef typename in_lno_row_view_t::const_value_type const_row_lno_t;
-  typedef typename in_lno_row_view_t::const_value_type const_size_type;
-  typedef typename in_lno_row_view_t::non_const_value_type non_const_row_lno_t;
-  typedef typename in_lno_row_view_t::const_type const_lno_row_view_t;
-  typedef typename in_lno_row_view_t::non_const_type non_const_lno_row_view_t;
-  typedef typename in_lno_nnz_view_t::const_value_type const_nnz_lno_t;
-  typedef typename in_lno_nnz_view_t::const_type const_lno_nnz_view_t;
-  typedef typename in_lno_nnz_view_t::non_const_type non_const_lno_nnz_view_t;
-  typedef typename in_scalar_nnz_view_t::const_data_type const_nnz_scalar_t; //nnz_scalar_t
-  typedef typename in_scalar_nnz_view_t::non_const_data_type non_const_nnz_scalar_t;
-  typedef typename in_scalar_nnz_view_t::const_type const_scalar_nnz_view_t;
-  typedef typename in_scalar_nnz_view_t::non_const_type non_const_scalar_nnz_view_t;
-  typedef typename KokkosGraph::GraphColoringHandle
-      <in_lno_row_view_t, non_const_lno_nnz_view_t, in_lno_nnz_view_t,
-      ExecutionSpace, TemporaryMemorySpace, PersistentMemorySpace> GraphColoringHandleType;
-  typedef typename KokkosSparse::GaussSeidelHandle
-      <in_lno_row_view_t, in_lno_nnz_view_t, in_scalar_nnz_view_t,
-      ExecutionSpace, TemporaryMemorySpace, PersistentMemorySpace> GaussSeidelHandleType;
+  typedef typename std::remove_const<lno_t_>::type  nnz_lno_t;
+  typedef const nnz_lno_t const_nnz_lno_t;
 
-  typedef typename KokkosSparse::SPGEMMHandle
-      <in_lno_row_view_t, in_lno_nnz_view_t, in_scalar_nnz_view_t,
-      ExecutionSpace, TemporaryMemorySpace, PersistentMemorySpace> SPGEMMHandleType;
+  typedef typename std::remove_const<scalar_t_>::type  nnz_scalar_t;
+  typedef const nnz_scalar_t const_nnz_scalar_t;
+
+  template <  typename right_size_type_, typename right_lno_t_, typename right_scalar_t_,
+              typename right_ExecutionSpace, typename right_TemporaryMemorySpace, typename right_PersistentMemorySpace>
+  //KokkosKernelsHandle<const_size_type,const_nnz_lno_t, const_nnz_scalar_t, HandleExecSpace, HandleTempMemorySpace, HandlePersistentMemorySpace> &operator=
+  KokkosKernelsHandle
+      (KokkosKernelsHandle<right_size_type_, right_lno_t_, right_scalar_t_,
+            right_ExecutionSpace, right_TemporaryMemorySpace, right_PersistentMemorySpace> & right_side_handle){
+
+    static_assert (std::is_same<size_type_,  const_size_type>::value,
+        "Kernel handle left hand side should have const size type in assignment");
+    static_assert (std::is_same<lno_t_,  const_nnz_lno_t>::value,
+        "Kernel handle left hand side should have const lno type in assignment");
+    static_assert (std::is_same<scalar_t_,  const_nnz_scalar_t>::value,
+        "Kernel handle left hand side should have const scalar type in assignment");
+
+    static_assert (std::is_same<ExecutionSpace,  HandleExecSpace>::value,
+        "Kernel handle left hand side should have execution space in assignment");
+    static_assert (std::is_same<TemporaryMemorySpace,  HandleTempMemorySpace>::value,
+        "Kernel handle left hand side should have temp memory space in assignment");
+    static_assert (std::is_same<PersistentMemorySpace,  HandlePersistentMemorySpace>::value,
+        "Kernel handle left hand side should have persistent memory space in assignment");
+
+    typedef typename  std::remove_const<right_size_type_>::type  nonconst_right_size_type;
+    typedef const nonconst_right_size_type const_right_size_type;
+
+    typedef typename  std::remove_const<right_lno_t_>::type  nonconst_right_lno_t;
+    typedef const nonconst_right_lno_t const_right_lno_t;
+
+    typedef typename  std::remove_const<right_scalar_t_>::type  nonconst_right_scalar_t;
+    typedef const nonconst_right_scalar_t const_right_scalar_t;
+
+
+    static_assert (std::is_same<size_type_,  const_right_size_type>::value,
+        "Kernel handle left and right sides should have same size type in assignment");
+    static_assert (std::is_same<lno_t_,  const_right_lno_t>::value,
+        "Kernel handle left and right sides should have same lno type in assignment");
+    static_assert (std::is_same<scalar_t_,  const_right_scalar_t>::value,
+        "Kernel handle left and right sides should have same scalar type in assignment");
+
+
+    static_assert (std::is_same<typename ExecutionSpace::execution_space,  typename right_ExecutionSpace::execution_space>::value,
+        "Kernel handle left and right sides should have same execution_space in assignment");
+    /*
+    static_assert (std::is_same<typename TemporaryMemorySpace::execution_space, typename right_TemporaryMemorySpace::execution_space>::value,
+        "Kernel handle left and right sides should have same TemporaryMemorySpace in assignment");
+    static_assert (std::is_same<typename PersistentMemorySpace::execution_space, typename right_PersistentMemorySpace::execution_space>::value,
+        "Kernel handle left and right sides should have same PersistentMemorySpace in assignment");
+      */
+    static_assert (std::is_same<typename ExecutionSpace::memory_space, typename right_ExecutionSpace::memory_space>::value,
+        "Kernel handle left and right sides should have same ExecutionSpace in assignment");
+    static_assert (std::is_same<typename TemporaryMemorySpace::memory_space, typename right_TemporaryMemorySpace::memory_space>::value,
+        "Kernel handle left and right sides should have same TemporaryMemorySpace in assignment");
+    static_assert (std::is_same<typename PersistentMemorySpace::memory_space, typename right_PersistentMemorySpace::memory_space>::value,
+        "Kernel handle left and right sides should have same PersistentMemorySpace in assignment");
+
+
+    this->gcHandle    = right_side_handle.get_graph_coloring_handle();
+    this->gcHandle_d2 = right_side_handle.get_distance2_graph_coloring_handle();
+
+    this->gsHandle = right_side_handle.get_gs_handle();
+    this->spgemmHandle = right_side_handle.get_spgemm_handle();
+
+    this->team_work_size = right_side_handle.get_set_team_work_size();
+    this->shared_memory_size = right_side_handle.get_shmem_size();
+    this->suggested_team_size = right_side_handle.get_set_suggested_team_size();
+
+    this->my_exec_space = right_side_handle.get_handle_exec_space();
+    this->use_dynamic_scheduling = right_side_handle.is_dynamic_scheduling();
+    this->KKVERBOSE = right_side_handle.get_verbose();
+    this->vector_size = right_side_handle.get_set_suggested_vector_size();
+
+    is_owner_of_the_gc_handle = false;
+    is_owner_of_the_d2_gc_handle = false;
+    is_owner_of_the_gs_handle = false;
+    is_owner_of_the_spgemm_handle = false;
+    is_owner_of_the_spadd_handle = false;
+    //return *this;
+  }
+
+
+  typedef typename KokkosGraph::
+    GraphColoringHandle<const_size_type, const_nnz_lno_t, const_nnz_lno_t, HandleExecSpace, HandleTempMemorySpace, HandlePersistentMemorySpace>
+      GraphColoringHandleType;
+
+  typedef typename KokkosGraph::
+    GraphColorDistance2Handle<const_size_type, const_nnz_lno_t, const_nnz_lno_t, HandleExecSpace, HandleTempMemorySpace, HandlePersistentMemorySpace>
+      GraphColorDistance2HandleType;
+
+  typedef typename KokkosSparse::
+    GaussSeidelHandle<const_size_type, const_nnz_lno_t, const_nnz_scalar_t, HandleExecSpace, HandleTempMemorySpace, HandlePersistentMemorySpace>
+      GaussSeidelHandleType;
+
+  typedef typename KokkosSparse::
+    SPGEMMHandle<const_size_type, const_nnz_lno_t, const_nnz_scalar_t, HandleExecSpace, HandleTempMemorySpace, HandlePersistentMemorySpace>
+      SPGEMMHandleType;
+
+  typedef typename Kokkos::View<nnz_scalar_t *, HandleTempMemorySpace> in_scalar_nnz_view_t;
 
   typedef typename Kokkos::View<size_type *, HandleTempMemorySpace> row_lno_temp_work_view_t;
   typedef typename Kokkos::View<size_type *, HandleTempMemorySpace> size_type_temp_work_view_t;
@@ -116,10 +186,20 @@ public:
   typedef typename Kokkos::View<bool *, HandlePersistentMemorySpace> bool_persistent_view_t;
   typedef typename Kokkos::View<bool *, HandleTempMemorySpace> bool_temp_view_t;
 
+  typedef
+    typename KokkosSparse::SPADDHandle<row_lno_temp_work_view_t, nnz_lno_temp_work_view_t, scalar_temp_work_view_t, HandleExecSpace, HandleTempMemorySpace>
+      SPADDHandleType;
+
+
 private:
+
   GraphColoringHandleType *gcHandle;
+  GraphColorDistance2HandleType *gcHandle_d2;
+
   GaussSeidelHandleType *gsHandle;
   SPGEMMHandleType *spgemmHandle;
+  SPADDHandleType *spaddHandle;
+
   int team_work_size;
   size_t shared_memory_size;
   int suggested_team_size;
@@ -128,21 +208,41 @@ private:
   bool use_dynamic_scheduling;
   bool KKVERBOSE;
   int vector_size;
+
+  bool is_owner_of_the_gc_handle;
+  bool is_owner_of_the_d2_gc_handle;
+  bool is_owner_of_the_gs_handle;
+  bool is_owner_of_the_spgemm_handle;
+  bool is_owner_of_the_spadd_handle;
+
 public:
 
-
-
-  KokkosKernelsHandle():
-      gcHandle(NULL), gsHandle(NULL),spgemmHandle(NULL),
-      team_work_size (-1), shared_memory_size(16128),
-      suggested_team_size(-1),
-      my_exec_space(KokkosKernels::Impl::kk_get_exec_space_type<HandleExecSpace>()),
-      use_dynamic_scheduling(true), KKVERBOSE(false),vector_size(-1){}
+  KokkosKernelsHandle()
+    : gcHandle(NULL)
+    , gcHandle_d2(NULL)
+    , gsHandle(NULL)
+    , spgemmHandle(NULL)
+    , spaddHandle(NULL)
+    , team_work_size(-1)
+    , shared_memory_size(16128)
+    , suggested_team_size(-1)
+    , my_exec_space(KokkosKernels::Impl::kk_get_exec_space_type<HandleExecSpace>())
+    , use_dynamic_scheduling(true)
+    , KKVERBOSE(false)
+    , vector_size(-1)
+    , is_owner_of_the_gc_handle(true)
+    , is_owner_of_the_d2_gc_handle(true)
+    , is_owner_of_the_gs_handle(true)
+    , is_owner_of_the_spgemm_handle(true)
+    , is_owner_of_the_spadd_handle(true)
+  { }
 
   ~KokkosKernelsHandle(){
     this->destroy_gs_handle();
     this->destroy_graph_coloring_handle();
+    this->destroy_distance2_graph_coloring_handle();
     this->destroy_spgemm_handle();
+    this->destroy_spadd_handle();
   }
 
 
@@ -168,7 +268,9 @@ public:
   void set_team_work_size(const int team_work_size_){
     this->team_work_size = team_work_size_;
   }
-
+  int get_set_team_work_size(){
+    return this->team_work_size;
+  }
   /**
    * \brief Returns the enum type for the execution space.
    */
@@ -184,7 +286,7 @@ public:
    * \param overall_work_size: The overall work size.
    */
   int get_team_work_size(const int team_size, const int concurrency, const nnz_lno_t overall_work_size){
-    if (this->team_work_size != -1){
+    if (this->team_work_size != -1) {
       return this->team_work_size;
     }
     else {
@@ -194,7 +296,6 @@ public:
       else {
         return 16;
       }
-
     }
   }
 
@@ -252,6 +353,9 @@ public:
     this->vector_size = vector_size_;
   }
 
+  int get_set_suggested_vector_size(){
+    return this->vector_size;
+  }
   /**
    * \brief Sets the team size to be used by the kernels. On GPUs and CPUs
    * usually the defaults are fine. But on CPUs with hyperthreads it might be
@@ -262,6 +366,9 @@ public:
     this->suggested_team_size = suggested_team_size_;
   }
 
+  int get_set_suggested_team_size(){
+    return this->suggested_team_size;
+  }
 
 
   /**
@@ -279,37 +386,83 @@ public:
 
 
 
-
+  // SPGEM
   SPGEMMHandleType *get_spgemm_handle(){
     return this->spgemmHandle;
   }
-
   void create_spgemm_handle(KokkosSparse::SPGEMMAlgorithm spgemm_algo = KokkosSparse::SPGEMM_DEFAULT){
     this->destroy_spgemm_handle();
+    this->is_owner_of_the_spgemm_handle = true;
     this->spgemmHandle = new SPGEMMHandleType(spgemm_algo);
-
   }
   void destroy_spgemm_handle(){
-    if (this->spgemmHandle != NULL){
+    if (is_owner_of_the_spgemm_handle && this->spgemmHandle != NULL){
       delete this->spgemmHandle;
       this->spgemmHandle = NULL;
     }
   }
 
+
+
+  // Distance-1 Graph Coloring
   GraphColoringHandleType *get_graph_coloring_handle(){
+    // (wcmclen): Should there be a check here to make sure we've created a GC handle before
+    //            handing the pointer out to something? This is disabled for now because it
+    //            gets thrown in tests run by spot check. Moving forward, we should consider
+    //            whether a "get the handle ptr, then allocate" vs. "only give out the handle ptr
+    //            if it actually exists" model.
+    //if(!this->is_owner_of_the_gc_handle)
+    //{
+    //  throw std::runtime_error("Graph coloring handle has not been created.");
+    //}
     return this->gcHandle;
   }
   void create_graph_coloring_handle(KokkosGraph::ColoringAlgorithm coloring_type = KokkosGraph::COLORING_DEFAULT){
     this->destroy_graph_coloring_handle();
+    this->is_owner_of_the_gc_handle = true;
     this->gcHandle = new GraphColoringHandleType();
     this->gcHandle->set_algorithm(coloring_type, true);
+    this->gcHandle->set_tictoc(KKVERBOSE);
   }
   void destroy_graph_coloring_handle(){
-    if (this->gcHandle != NULL){
+    if (is_owner_of_the_gc_handle &&  this->gcHandle != NULL){
       delete this->gcHandle;
       this->gcHandle = NULL;
     }
   }
+
+
+
+  // Distance-2 Graph Coloring
+  GraphColorDistance2HandleType *get_distance2_graph_coloring_handle()
+  {
+    /* disabled for consistency with `get_graph_coloring_handle()`. See the comment there
+       for reasons.
+    if(!this->is_owner_of_the_d2_gc_handle)
+    {
+      throw std::runtime_error("D2 graph coloring handle has not been created.");
+    }
+    */
+    return this->gcHandle_d2;
+  }
+  void create_distance2_graph_coloring_handle(KokkosGraph::GraphColoringAlgorithmDistance2 coloring_type = KokkosGraph::COLORING_D2_DEFAULT)
+  {
+    this->destroy_distance2_graph_coloring_handle();
+    this->is_owner_of_the_d2_gc_handle = true;
+    this->gcHandle_d2 = new GraphColorDistance2HandleType();
+    this->gcHandle_d2->set_algorithm(coloring_type, true);
+    this->gcHandle_d2->set_tictoc(KKVERBOSE);
+    this->gcHandle_d2->set_verbose(KKVERBOSE);
+  }
+  void destroy_distance2_graph_coloring_handle()
+  {
+    if(is_owner_of_the_d2_gc_handle && this->gcHandle_d2 != NULL)
+    {
+      delete this->gcHandle_d2;
+      this->gcHandle_d2 = NULL;
+    }
+  }
+
 
 
   GaussSeidelHandleType *get_gs_handle(){
@@ -318,10 +471,11 @@ public:
   void create_gs_handle(
     KokkosSparse::GSAlgorithm gs_algorithm = KokkosSparse::GS_DEFAULT){
     this->destroy_gs_handle();
+    this->is_owner_of_the_gs_handle = true;
     this->gsHandle = new GaussSeidelHandleType(gs_algorithm);
   }
   void destroy_gs_handle(){
-    if (this->gsHandle != NULL){
+    if (is_owner_of_the_gs_handle && this->gsHandle != NULL){
       if (this->gsHandle->is_owner_of_coloring()){
         this->destroy_graph_coloring_handle();
       }
@@ -332,7 +486,24 @@ public:
 
 
 
-};
+  SPADDHandleType *get_spadd_handle(){
+    return this->spaddHandle;
+  }
+  void create_spadd_handle(bool input_sorted) {
+    this->destroy_spadd_handle();
+    this->is_owner_of_the_spadd_handle = true;
+    this->spaddHandle = new SPADDHandleType(input_sorted);
+  }
+  void destroy_spadd_handle(){
+    if (is_owner_of_the_spadd_handle && this->spaddHandle != NULL)
+    {
+      delete this->spaddHandle;
+      this->spaddHandle = NULL;
+    }
+  }
+
+
+};    // end class KokkosKernelsHandle
 
 }
 }
